TD-277: decouple backup & restore use cases and endpoints from schema
The backup and restore use cases are completely implemented by the package usecases/backup

Backup and restore endpoints are handled separately from schema endpoints
