+++ usecases/schema/manager.go
@@ -21,7 +21,6 @@ import (
 	"github.com/semi-technologies/weaviate/entities/schema"
 	"github.com/semi-technologies/weaviate/usecases/cluster"
 	"github.com/semi-technologies/weaviate/usecases/config"
-	"github.com/semi-technologies/weaviate/usecases/schema/backups"
 	"github.com/semi-technologies/weaviate/usecases/schema/migrate"
 	"github.com/semi-technologies/weaviate/usecases/sharding"
 	"github.com/sirupsen/logrus"
@@ -43,7 +42,6 @@ type Manager struct {
 	clusterState            clusterState
 	hnswConfigParser        VectorConfigParser
 	invertedConfigValidator InvertedConfigValidator
-	backups                 backups.BackupManager
 	RestoreStatus           sync.Map
 	RestoreError            sync.Map
 	sync.Mutex
@@ -94,7 +92,7 @@ func NewManager(migrator migrate.Migrator, repo Repo,
 	hnswConfigParser VectorConfigParser, vectorizerValidator VectorizerValidator,
 	invertedConfigValidator InvertedConfigValidator,
 	moduleConfig ModuleConfig, clusterState clusterState,
-	txClient cluster.Client, backupManager backups.BackupManager,
+	txClient cluster.Client,
 ) (*Manager, error) {
 	m := &Manager{
 		config:                  config,
@@ -109,7 +107,6 @@ func NewManager(migrator migrate.Migrator, repo Repo,
 		moduleConfig:            moduleConfig,
 		cluster:                 cluster.NewTxManager(cluster.NewTxBroadcaster(clusterState, txClient)),
 		clusterState:            clusterState,
-		backups:                 backupManager,
 	}
 
 	m.cluster.SetCommitFn(m.handleCommit)
